Add tracing of error recovery
Problem: sometimes we need some debug instrumentation to understand what is going on.
For example which candidates were considered by recovery algorithm, which symbols
were on parser stack in that moment and so on.

Solution: extend interface of the `Merlin_recovery` module by `PRINTER` that
encapsulate callback for printing symbols, elements, envs and so on. And also
provide functor `MakePrinter` that instantiate `PRINTER` by using
minimal subset of the callbacks and `MenhirLib.Printer.Make`.
Demo driver also changed.