[LIGO-329] Emitter bugfix
Problem:
There is a difference between Merlin and ocaml-recovery-parser in the
 emitter.ml which leads to poor results on some tests.

Explanation:
There is a `Sub` action that forces the error-recovery algorithm to
apply all recovery steps at once without trying to recover in the
middle of it. The difference is that ocaml-recovery-parser uses `Sub`
more often than Merlin. Here is the example.
The test is missing the `with` keyword.
`MATCH expr . WITH ...`

ocaml-recovery-parser
```
...
let r60 = Sub [S (T T_With); r59] in
...
| 652 -> One [r60]
...
```
Merlin
```
...
let r371 = S (T T_WITH) :: r370 in
...
| 1147 -> One (r371)
...
```
This patch
```
...
let r60 = S (T T_With) :: r59 in
...
| 652 -> One (r60)
...
```
As you can see, the recovery algorithm able to shift `with` and
finish recovery only in the last two cases while in the first
case algorithm have to apply all recovery steps which lead to poor
results on this test

Solution:
Change emitter.ml corresponding to Merlin's