fix opam package license
Problem: when opam installs this package (e.g. from pin-dependency) it throws:
"warning 62: License doesn't adhere to the SPDX standard..."
The reason is that following SPDX standard (Annex SPDX License expression)
"operators (AND, OR, WITH) should be matched in case-sensitive manner"

Solution: fix case of and-operator in the opam file.