[LIGO-407] make cost model more flexible
Problem: in the current cost model default costs of shift and reduce action is
the same but in some cases a reduction is a mere way to lift parser up to a next
level precedence.

Solution: make default cost of each action is customizable by new grammar-level
attributes: `recover.default_cost_of_prod`, `recover.default_cost_of_symbol`.

Also, cost models is changed a bit:
- before: cost of symbol equals 1 + cost from its attributes
- after:  cost of symbol equals whether cost from its attributes if we have any
of them or otherwise default cost of symbol (which
`recover.default_cost_of_symbol` or 1 if it is missed)

Other improvements:
- replace string literals with string constants for attribute names
- update cost report