[LIGO-458] Replace the old precalculation algorithm with the new one
Problem: The old precalculation algorithm has at least two major bugs:
1) The old algorithm implies a kind of Dijkstra algorithm but the state
sorting is based on nonterminal id rather than the path cost. It luckily
works in many cases due to the graph simplicity, but in other cases, it
finds not the cheapest path which leads to unpredictable recovery mistakes
2) The old algorithm wrongly handles the cases where the parent is the
initial vertex. It can't find any path in such cases and just does a random
step. It luckily works sometimes, but sometimes it leads to unpredictable
recovery mistakes or even to looping in one case in JsLIGO.

Solution: The new algorithm normally handles the cases where the parent is
the initial vertex and has a priority queue for a proper sorting in the
Dijkstra algorithm